Rename the facades
We hacked together the aptly named `Facade2` during LN squad.  This
patch renames it to `Facade` after first renaming the original facade to
be `Rfc003Facade` in line with changes made previously to highlight what
code is rfc003.
