+++ cnd/src/http_api/route_factory.rs
@@ -2,7 +2,7 @@ use crate::{
     config::settings::AllowedOrigins,
     http_api,
     network::LocalPeerId,
-    swap_protocols::{self, rfc003::SwapId, Facade, Facade2, LocalSwapId},
+    swap_protocols::{self, rfc003::SwapId, Facade, LocalSwapId, Rfc003Facade},
 };
 use warp::{self, filters::BoxedFilter, Filter, Reply};
 
@@ -17,17 +17,17 @@ pub fn new_action_link(id: &SwapId, action: &str) -> String {
 }
 
 pub fn create(
-    dependencies: Facade,
-    facade2: Facade2,
+    rfc003_facade: Rfc003Facade,
+    facade: Facade,
     allowed_origins: &AllowedOrigins,
 ) -> BoxedFilter<(impl Reply,)> {
-    let peer_id = dependencies.local_peer_id();
+    let peer_id = rfc003_facade.local_peer_id();
     let swaps = warp::path(http_api::PATH);
     let rfc003 = swaps.and(warp::path(RFC003));
     let peer_id = warp::any().map(move || peer_id.clone());
     let empty_json_body = warp::any().map(|| serde_json::json!({}));
-    let dependencies = warp::any().map(move || dependencies.clone());
-    let facade2 = warp::any().map(move || facade2.clone());
+    let rfc003_facade = warp::any().map(move || rfc003_facade.clone());
+    let facade = warp::any().map(move || facade.clone());
 
     let cors = warp::cors()
         .allow_methods(vec!["GET", "POST"])
@@ -45,13 +45,13 @@ pub fn create(
     let rfc003_post_swap = rfc003
         .and(warp::path::end())
         .and(warp::post())
-        .and(dependencies.clone())
+        .and(rfc003_facade.clone())
         .and(warp::body::json())
         .and_then(http_api::routes::rfc003::post_swap);
 
     let rfc003_get_swap = rfc003
         .and(warp::get())
-        .and(dependencies.clone())
+        .and(rfc003_facade.clone())
         .and(warp::path::param())
         .and(warp::path::end())
         .and_then(http_api::routes::rfc003::get_swap);
@@ -59,7 +59,7 @@ pub fn create(
     let get_swaps = swaps
         .and(warp::get())
         .and(warp::path::end())
-        .and(dependencies.clone())
+        .and(rfc003_facade.clone())
         .and_then(http_api::routes::index::get_swaps);
 
     let rfc003_action = warp::method()
@@ -70,27 +70,27 @@ pub fn create(
         >())
         .and(warp::path::end())
         .and(warp::query::<http_api::action::ActionExecutionParameters>())
-        .and(dependencies.clone())
+        .and(rfc003_facade.clone())
         .and(warp::body::json().or(empty_json_body).unify())
         .and_then(http_api::routes::rfc003::action);
 
     let get_peers = warp::get()
         .and(warp::path("peers"))
         .and(warp::path::end())
-        .and(dependencies.clone())
+        .and(rfc003_facade.clone())
         .and_then(http_api::routes::peers::get_peers);
 
     let get_info_siren = warp::get()
         .and(warp::path::end())
         .and(warp::header::exact("accept", "application/vnd.siren+json"))
         .and(peer_id.clone())
-        .and(dependencies.clone())
+        .and(rfc003_facade.clone())
         .and_then(http_api::routes::index::get_info_siren);
 
     let get_info = warp::get()
         .and(warp::path::end())
         .and(peer_id)
-        .and(dependencies)
+        .and(rfc003_facade)
         .and_then(http_api::routes::index::get_info);
 
     let han_ethereum_halight_bitcoin = warp::post()
@@ -99,7 +99,7 @@ pub fn create(
         ))
         .and(warp::path::end())
         .and(warp::body::json())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::index::post_han_ethereum_halight_bitcoin);
 
     let herc20_halight_bitcoin = warp::post()
@@ -108,7 +108,7 @@ pub fn create(
         ))
         .and(warp::path::end())
         .and(warp::body::json())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::index::post_herc20_halight_bitcoin);
 
     let halight_bitcoin_han_ether = warp::post()
@@ -117,7 +117,7 @@ pub fn create(
         ))
         .and(warp::path::end())
         .and(warp::body::json())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::index::post_halight_bitcoin_han_ether);
 
     let halight_bitcoin_herc20 = warp::post()
@@ -126,14 +126,14 @@ pub fn create(
         ))
         .and(warp::path::end())
         .and(warp::body::json())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::index::post_halight_bitcoin_herc20);
 
     let get_halight_swap = swaps
         .and(warp::get())
         .and(warp::path::param())
         .and(warp::path::end())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::get_halight_swap);
 
     let lightning_action_init = swaps
@@ -141,7 +141,7 @@ pub fn create(
         .and(warp::path::param::<LocalSwapId>())
         .and(warp::path("init"))
         .and(warp::path::end())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::action_init);
 
     let lightning_action_fund = swaps
@@ -149,7 +149,7 @@ pub fn create(
         .and(warp::path::param::<LocalSwapId>())
         .and(warp::path("fund"))
         .and(warp::path::end())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::action_fund);
 
     let lightning_action_redeem = swaps
@@ -157,7 +157,7 @@ pub fn create(
         .and(warp::path::param::<LocalSwapId>())
         .and(warp::path("redeem"))
         .and(warp::path::end())
-        .and(facade2.clone())
+        .and(facade.clone())
         .and_then(http_api::routes::action_redeem);
 
     let lightning_action_refund = swaps
@@ -165,7 +165,7 @@ pub fn create(
         .and(warp::path::param::<LocalSwapId>())
         .and(warp::path("refund"))
         .and(warp::path::end())
-        .and(facade2)
+        .and(facade)
         .and_then(http_api::routes::action_refund);
 
     preflight_cors_route
