+++ cnd/src/http_api/routes/rfc003/handlers/get_swap.rs
@@ -1,10 +1,13 @@
 use crate::{
     db::{DetermineTypes, Retrieve},
     http_api::swap_resource::{build_rfc003_siren_entity, IncludeState, OnFail},
-    swap_protocols::{rfc003::SwapId, Facade},
+    swap_protocols::{rfc003::SwapId, Rfc003Facade},
 };
 
-pub async fn handle_get_swap(dependencies: Facade, id: SwapId) -> anyhow::Result<siren::Entity> {
+pub async fn handle_get_swap(
+    dependencies: Rfc003Facade,
+    id: SwapId,
+) -> anyhow::Result<siren::Entity> {
     let swap = Retrieve::get(&dependencies, &id).await?;
     let types = dependencies.determine_types(&id).await?;
