+++ cnd/src/init_swap.rs
@@ -7,18 +7,18 @@ use crate::{
             state::Insert,
             Accept, Request, SwapCommunication,
         },
-        Facade,
+        Rfc003Facade,
     },
 };
 use tracing_futures::Instrument;
 
 #[allow(clippy::cognitive_complexity)]
 pub async fn init_accepted_swap<AL, BL, AA, BA, AH, BH, AI, BI, AT, BT>(
-    dependencies: &Facade,
+    dependencies: &Rfc003Facade,
     accepted: AcceptedSwap<AL, BL, AA, BA, AI, BI>,
 ) -> anyhow::Result<()>
 where
-    Facade: HtlcFunded<AL, AA, AH, AI, AT>
+    Rfc003Facade: HtlcFunded<AL, AA, AH, AI, AT>
         + HtlcFunded<BL, BA, BH, BI, BT>
         + HtlcDeployed<AL, AA, AH, AI, AT>
         + HtlcDeployed<BL, BA, BH, BI, BT>
