+++ cnd/src/load_swaps.rs
@@ -2,11 +2,11 @@
 use crate::{
     db::{DetermineTypes, LoadAcceptedSwap, Retrieve},
     init_swap::init_accepted_swap,
-    swap_protocols::Facade,
+    swap_protocols::Rfc003Facade,
 };
 
 #[allow(clippy::cognitive_complexity)]
-pub async fn load_swaps_from_database(facade: Facade) -> anyhow::Result<()> {
+pub async fn load_swaps_from_database(facade: Rfc003Facade) -> anyhow::Result<()> {
     tracing::debug!("loading swaps from database ...");
 
     for swap in Retrieve::all(&facade).await?.iter() {
